+++ frameworks/ans/core/include/ans_manager_stub.h
@@ -74,7 +74,7 @@ public:
      * @param label Indicates the label of the notification to cancel.
      * @return Returns cancel notification result.
      */
-    virtual ErrCode Cancel(int notificationId, const std::string &label) override;
+    virtual ErrCode Cancel(int32_t notificationId, const std::string &label) override;
 
     /**
      * @brief Cancels all the published notifications.
@@ -168,7 +168,7 @@ public:
      * @param num Indicates the number of slot.
      * @return Returns ERR_OK on success, others on failure.
      */
-    virtual ErrCode GetSlotNumAsBundle(const sptr<NotificationBundleOption> &bundleOption, int &num) override;
+    virtual ErrCode GetSlotNumAsBundle(const sptr<NotificationBundleOption> &bundleOption, uint64_t &num) override;
 
     /**
      * @brief Deletes multiple notification slot groups.
@@ -193,7 +193,7 @@ public:
      * @param num Indicates the number of active notifications of the current application.
      * @return Returns ERR_OK on success, others on failure.
      */
-    virtual ErrCode GetActiveNotificationNums(int &num) override;
+    virtual ErrCode GetActiveNotificationNums(uint64_t &num) override;
 
     /**
      * @brief Obtains all active notifications in the current system. The caller must have system permissions to
@@ -267,7 +267,7 @@ public:
      * @param num Indicates the badge number.
      * @return Returns ERR_OK on success, others on failure.
      */
-    virtual ErrCode SetNotificationBadgeNum(int num) override;
+    virtual ErrCode SetNotificationBadgeNum(int32_t num) override;
 
     /**
      * @brief Obtains the importance level of this application.
@@ -276,7 +276,7 @@ public:
                LEVEL_MIN, LEVEL_LOW, LEVEL_DEFAULT, LEVEL_HIGH, or LEVEL_UNDEFINED.
      * @return Returns ERR_OK on success, others on failure.
      */
-    virtual ErrCode GetBundleImportance(int &importance) override;
+    virtual ErrCode GetBundleImportance(int32_t &importance) override;
 
     /**
      * @brief Checks whether this application has permission to modify the Do Not Disturb (DND) notification policy.
@@ -311,7 +311,7 @@ public:
      * @return Returns ERR_OK on success, others on failure.
      */
     virtual ErrCode RemoveNotification(
-        const sptr<NotificationBundleOption> &bundleOption, int notificationId, const std::string &label) override;
+        const sptr<NotificationBundleOption> &bundleOption, int32_t notificationId, const std::string &label) override;
 
     /**
      * @brief Delete all notifications.
